Fix bug breaking conversion of records to a df
When only list columns were generated
as.data.frame.finbif_occ would fail to intialise
a data.frame to which to add the list columns to.
The fix was to add a temporary column just before
initialising the data.frame ensuring there is a
data.frame to add list columns to.