Switch vignettes to rmarkdown/move figures to man
Switching the vignettes type to html_vignette and
building the vignettes with knitr::rmarkdown
instead of knitr::knitr fixes the problem of the
vignette front matter not showing up in the R help
browser. To avoid failing vignette rebuild in R
CMD check, need to move figures to man/figures
instead of vignettes/figure.