Reduce no. of default vars for occurrence records
This also necessitates a change in the way
date_time is calculated when calling
finbif_occurrence. When date_time is TRUE
then the variables neccessary to compute it are
requested regardless of user input. Then only
those variables specifically requested by the user
are returned in the resulting finbif_occ object.