Add support for localised data
On loading, if unset by the user, the locale is
guessed from the system settings and stored in the
package option "finbif_locale". For data where
more than one language is available (e.g.,
vernacular names) the value of "finbif_locale"
will be preferred, falling back to other supported
languages in a set order. The default language is
English. Locale can also be set per-record
request.