Add ability to get complex variables from records
For example the variables associated with record
media (e.g., url, type, etc.) were not previously
accessible as the internal function
`get_el_recurse` could not find them. Now if the
recursion encounters a level with no name it
checks if the level below has names and proceeds
if it can find a collection of elements with the
given name.