Fix bug triggered when selecting certain vars
Extracting variables from finbif_records obj
would fail if a single value variable was
missing but shared the same suffix as another
variable. In such cases it appeared to be a
list var. The solution is to detect list vars
by the absence of 'names' rather than the
presence of 'names' in their elements.