Improve the way 'municipality' data is handled
What a 'municipality' is, varies from country to
country. Strictly speaking, a Finnish
'municipality', in the Darwin Core standard, is
a 'county'. But in the USA for example a
'municipality' is a level of organisation below
the county level.

These changes account for these
differences by adding a level below 'municipality'
in the Finnish sense which is called 'local_area'
(or 'municipality' in DwC). It is assumed that
non-Finnish data will use the terms 'county' and
and 'municipality' in the raw data (e.g., data
coming from the Kotka CMS) in the sense they are
outline in Darwin Core. And Finnish data will map
'municipality' to DwC 'county' lacking any true
'municipality' data.